Moved state around for checkout
